One dialog to rule them all
