+++ demo/src/main/java/com/jrummyapps/android/colorpicker/demo/ColorPickerActivity.java
@@ -58,6 +58,10 @@ public class ColorPickerActivity extends Activity implements ColorPickerView.OnC
     colorPickerView.setColor(initialColor, true);
     colorPanelView.setColor(initialColor);
 
+    // TODO: fix attributes not applying when inflating layout
+    colorPanelView.setShape(ColorPanelView.Shape.RECT);
+    newColorPanelView.setShape(ColorPanelView.Shape.RECT);
+
     btnOK.setOnClickListener(this);
     btnCancel.setOnClickListener(this);
